Better handling of atoms.info sanitizing
Sometimes the atoms.info dictionary can contain objects that are not JSON serializable (e.g. the Spacegroup object). This can lead to problems when storing the atoms.info data in a DB. 

This commit is a more robust implementation that should ensure any and all atoms.info data can be properly serialized.