Address breaking change with NewtonNet 1.1
NewtonNet 1.1 broke the energy and vibrational frequency for one of the unit tests. This PR uncomments that test to show precisely where the breaking change is, and before merging we need to fix it so the tests pass. Ideally, we should identify what changed in NewtonNet 1.1 that caused the energy/forces to change for this test.

This is a new PR to make things clearer, @kumaranu.