Remove calculator params from docstrings (#1232)
In principle, the idea was nice in that it would enable editors to show
the default params by default. In practice, it's messy because things
can accidentally become out of sync between the docstring and the code,
and sometimes it's not practical to list all parameters in the docstring
itself.