Add espresso recipes and custom calculator (#1238)
- [x] I have read the [Developer
Guide](https://quantum-accelerators.github.io/quacc/dev/contributing.html).
Don't lie! 😉
- [x] My PR is on a custom branch and is _not_ named `main`.
- [x] I have added relevant unit tests. Note: Your PR will likely not be
merged without proper tests.

A first PR for espresso additions, this is very much draft stage and
does not aim to work, just to show the general idea. Right now I have
other things to do but I believe this is a good first draft. There will
be probably a thousand details that you don't like in this PR (variable
names, etc...) so I will let you have a look. I will come back to it
later on.

The aim is to make it simple for the future when adding other espresso
binaries, for that I created mini-classes based on a more generic
'EspressoCalculator' class. The process to add a binary would be as
follows:

1. Add a class in quacc.calculators.espresso.calculator (few lines)
2. Add a Fortran keylist in quacc.calculators.espresso.keys
3. add a write_BIN_specific function into quacc.calculators.espresso.io
as well as read_bin_output

And that's it, there is nothing else to change, everything is done under
the hood. Stage 3 is by far the most time-consuming, 1 and 2 shouldn't
take more than a few minutes of work, and GPT is able to provide 2 if
needed based on the binary doc page.

Presets are a good idea, I added a quick one based on the cutoff table
given by the SSSP library. I am thinking about adding more specific ones
in the future (slabs, bulk...)

I think for the pp we shouldn't deal with env variables other than the
QUACC ones since I am not aware of how different workflow managers deal
with them. Covalent for example always uses env variables defined when
the server starts and it does not seem intuitive to change them on the
fly.

ASR edit: Closes #1224. Closes #1306. Closes #1307.

---------

Co-authored-by: Andrew S. Rosen <asrosen93@gmail.com>