Q-Chem recipe redesign (#1123)
This is a general overhaul of the Q-Chem recipes. There are many reasons
why a redesign is necessary. From a usability perspective, the recipes
are largely inconsistent with other recipes in `quacc`. Some input
parameters of interest are also entirely inaccessible by the recipes,
limiting flexibility for the user. It also relies heavily on knowledge
of the Pymatgen `QchemDictSet` kwargs, which I consider undesirable
(although we should support it of course).

This PR should resolve all those issues (and closes #1112) while still
retaining all the features of the original recipes.

Edit: Alright, this is done!! But I am holding off on merging this until
I run some production tests. If you're reading this and find this PR to
be important to you, feel free to ping me.

---------

Co-authored-by: deepsource-autofix[bot] <62050782+deepsource-autofix[bot]@users.noreply.github.com>