Use partial functions in flows to increase user flexibility in setting job attributes (e.g. executor) (#1322)
Closes #1320, although I'm not 100% confident that this is the ideal
approach yet. I'm not comfortable with merging this without further
justification that this is the best mechanism.

Pros:
- There is now full flexibility in setting the attributes of each job in
a workflow, meaning you can customize each executor
- It is clearer which kwargs are being set since they are associated
with the given function
- The user can even define their own custom job for a flow without
needing to create a new flow altogether

Cons:
- It is now clunkier to modify the kwargs of a job in a workflow because
you need to import the underlying job and make a partial function
- The user (likely) has to learn about partial functions from the docs
since it is a new concept to most people
- Cannot be used with Prefect since the kwarg is not JSON serialiazable.
See https://github.com/PrefectHQ/prefect/issues/10135.