Espresso: managing various outdirs (#1335)
## Required Checklist

- [x] I have read the [Contributing
Guide](https://quantum-accelerators.github.io/quacc/dev/contributing.html).
Don't lie! 😉
- [x] My PR is on a custom branch and is _not_ named `main`.
- [x] I have added relevant unit tests. Note: Your PR will likely not be
merged without proper tests.

If you are an external contributor, you will see a comment from
@buildbot-princeton. You can ignore this, as it is solely for admin
purposes.

## Summary of Changes

Edit: No need to be that complicated, this PR just makes sure that
chosen outdirs exist if inside the calc_dir.

If the user decides to ask for outdirs that are outside the calc_dir via
absolute paths, then there is nothing we should do (?). It is the user
responsibility to deal with that data (this PR add a dictionary to keep
track of that data but does nothing with it).

---------

Co-authored-by: Andrew S. Rosen <asrosen93@gmail.com>