Refactor espresso calculator and recipes (#1426)
#Summary

This PR seeks to streamline a few things related to the Espresso
calculator and recipes.

1. I merged in the changes from #1415, which moves the `phonon_job` to
its own module and makes the previous directory a required argument to
better inform the user what is needed.

2. I removed the `calc_defaults` kwarg from the calculator. This
shouldn't have been in there, as the way quacc works is that the
`calc_defaults` are handled by the `_base.py` functions, not in the
calculator itself. The functionality remains the same, but the logic for
merging the keyword arguments is done _before_ instantiating the
calculator.

3. Several aspects of the calculator were internally refactored without
any change of functionality to the recipes. For instance, see the
rewrite of `parse_pw_preset` to `parse_pseudopotential_info`.

4. There were a few instances before where the user-specified parameters
were overwritten by a preset, which is a strict no-no. We should always
respect the user's parameters. This was most notable with things like
`ecutwfc`. If the user specified a value of `50`, but the preset found a
maximum elemental `ecutwfc` of 100, it'd overwrite the user's `50`
value. In practice, if the user asks for a value, they should see that
value in the input file. In practice, if a user wants to rely on the max
value from the preset (typically the case), they should not explicitly
specify an `ecutwfc` at all.

5. The most notable change is that all parameters must be specified
explicitly with respect to a given section in the input file. For
instance, use `input_data = {"system": {"ecutwfc": 100}}` and not
`input_data={"ecutfc": 100}`. There are several important reasons for
this. The first is that ASE as a platform is trying to move away from
any enumeration of input arguments since it is difficult to keep up with
version updates, and we should do the same here. The second more
important reason is that the existing approach makes it non-trivial to
ensure the user-specified settings take priority.

For instance, consider a user-supplied argument of
`input_data={"system": {"ecutfc": 50}}` and a preset containing

```yaml
input_data:
  ecutwfc: 100
```

You can't merge the resulting dictionaries because then you'd end up
with `input_data = {"ecutwfc": 50, "system": {"ecutwfc": 100}}`. You can
get around this by converting _both_ the list of keyword arguments and
the presets to the same format with the list of acceptable keys, but
it's messy. In general, I prefer to keep things as 1:1 comparable with
the actual input file as possible, so it's easier to just remove the
support for these kinds of shortcuts.