Implement a minimal preset for Espresso (#1427)
Closes #1412. Also, I removed default parameters like `ecutwfc` and
`ecutrho`, which should not be in the recipe defaults since they would
be the same defaults for every recipe. The recipe defaults should have
the minimal info needed to run a reliable calculation for that
calculation type; we can put a default preset ("minimal.yaml") that will
automatically set `ecutwfc` and `ecutrho` to the max value based on the
pseudopotential YAML instead.

---------

Co-authored-by: deepsource-autofix[bot] <62050782+deepsource-autofix[bot]@users.noreply.github.com>