Avoid surprise deleting dense k-spacing from basic_pw preset
Situation: A user is used to the ASE espresso calculator and wants to run a very large isolated system in a big box. The user doesn't set any kpts because they are used to Gamma being default. Instead quacc set a very high kspacing, which in this case will ruin the calculation.

This patch aim to avoid that.