Use a (mostly) consistent default ASE optimizer (#1712)
## Summary of Changes

Sets a standardized, default ASE optimizers for ASE relaxations.

@Nekkrad, @tomdemeyere --- I saw that you both had LBFGS as the default
in Espresso and ONETEP. Are memory issues ever really a problem with
relaxations that would warrant its use? Right now, this PR has
everything switching to BFGS by default, but I'm not opposed to LBFGS if
there is sufficient motivation to do so. Interested to hear your
thoughts. Presumably extremely large systems it might matter? If so,
happy to use LBFGS here.

---------

Co-authored-by: deepsource-autofix[bot] <62050782+deepsource-autofix[bot]@users.noreply.github.com>