Refactor and streamline `copy_files` handling (#1759)
Closes #1742.

Note: 
- https://github.com/Parsl/parsl/issues/3108

## Motivation

In short, the `copy_files` handling was in desperate need of a
behind-the-scenes overhaul. There are several reasons for this.

1. As noted in #1742, you can't do `Path()` on a `Future`, which means
any recipes doing `Path()` (or string concatenation of paths) in a
`Flow` are going to crash when dispatched over a workflow engine.
2. There are several file handling functions that do redundant things or
are handled in existing dependencies: `copy_decompress_files`,
`copy_decompress_tree`, `copy_decompress_files_from_dir`.
3. Because of Point 2, not all the functions handle things in the same
way. For instance, only some have glob support right now, which is
confusing.
4. There are many instances where the type hints are simply inaccurate.
For instance, in the Espresso recipes, the `prev_dir: str | Path`
positional argument is often getting `dict` entries thrown into, as a
result of `pw_copy_files` and such (see the test suite).

Because of this mess, things were in desperate need of a revamp, which
this PR addresses.

## User-Facing Changes

In most cases, there are no breaking changes to users and is mostly an
internal refactor. Some positional arguments of `prev_dir` have been
renamed `copy_files` for consistency with the type hint, although no
user changes are expected if used as positional. Some internal functions
have slightly modified behavior, such as `copy_decompress_files`. Refer
to the function signature for more.

---------

Co-authored-by: deepsource-autofix[bot] <62050782+deepsource-autofix[bot]@users.noreply.github.com>