Increase hackability of recipes (#2017)
## Summary of Changes

This PR increases the "hackability" of recipes for users that are doing
funky things with a custom calculator. Namely, some recipes don't take a
`**calc_kwargs` because, well, there is no need to. An example of this
is the ORCA recipes since the ASE `ORCA` calculator takes
`orcasimpleinput` and `orcablocks` as dedicated keyword arguments; there
is no need for the user to pass an arbitrary set of `**calc_kwargs`
because the possible options are clearly enumerated in ASE. However, if
someone makes a custom `ORCA` calculator where they add on additional
keyword arguments, then they can't get passed anywhere. This is likely a
rare edge case, but it's pretty painless to implement.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>