Add support for dynamically changing settings with an activated workflow engine (v2) (#2277)
This PR is a collaboration with @honghuikim. It is a continuation of
#2163 that addresses changes introduced in #2276. For additional history
on this PR, please refer to #2163.

The focus of this PR is to enable dynamic changing of Pydantic model
settings when using an active workflow engine. This is achieved, in
large part, through a new `settings_swap` keyword argument to the
decorator.

Closes #2147.

---------

Co-authored-by: Honghui Kim <itab1107@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Honghui Kim <165984476+honghuikim@users.noreply.github.com>