Add a setting to store the logger file to disk (#2436)
## Summary of Changes

Provides a setting to store the log files to disk. Before, it was only
shown in the console. Now, we have an option to write it out to a custom
file if desired by the user.

@tomdemeyere: I think this explains part of the confusion. I had assumed
the logger metadata was being written out to disk and was confused why
you had to go digging through folders to find out which failed. However,
it was only shown in the console.

### Requirements

- [X] My PR is focused on a [single feature addition or
bugfix](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/getting-started/best-practices-for-pull-requests#write-small-prs).
- [X] My PR has relevant, comprehensive [unit
tests](https://quantum-accelerators.github.io/quacc/dev/contributing.html#unit-tests).
- [X] My PR is on a [custom
branch](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/proposing-changes-to-your-work-with-pull-requests/creating-and-deleting-branches-within-your-repository)
(i.e. is _not_ named `main`).

Note: If you are an external contributor, you will see a comment from
[@buildbot-princeton](https://github.com/buildbot-princeton). This is
solely for the maintainers.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>