fix doc of espresso.py (#2485)
Since it's relevant to Quantum Espresso, the "INCAR" used for VASP was
removed to avoid confusion.
Feel free to reject this pull request because it's just a small fix of
the document.

## Summary of Changes

Since it's relevant to Quantum Espresso, the "INCAR" used for VASP was
removed to avoid confusion.

### Requirements

- [ ] My PR is focused on a [single feature addition or
bugfix](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/getting-started/best-practices-for-pull-requests#write-small-prs).
- [ ] My PR has relevant, comprehensive [unit
tests](https://quantum-accelerators.github.io/quacc/dev/contributing.html#unit-tests).
- [ ] My PR is on a [custom
branch](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/proposing-changes-to-your-work-with-pull-requests/creating-and-deleting-branches-within-your-repository)
(i.e. is _not_ named `main`).

Note: If you are an external contributor, you will see a comment from
[@buildbot-princeton](https://github.com/buildbot-princeton). This is
solely for the maintainers.