Added ase Nudged Elastic Band in the NewtonNet recipe (#2176)
## Summary of Changes

\>>
Target here is to add ase's NEB as the double ended transition state
search method for NewtonNet recipe.

- I need to make sure that geodesic code is added as a preprocessor for
the double ended transition state search.
- NEB will take the output of Geodesic
- Sella will them run from the top point of NEB.
- Tests will need to be added for all the files keeping the coverage of
more than 99% before taking it for review.
- Currently, there is a lot of unstructured code in the ts.py and in the
test. This will need to be cleaned up to match with the rest of the
recipe's in QuAcc.
- The summary dictionaries will need to be written so that the data can
be passed back to the database.
- Also, the requirements will be added in the NewtonNet's recipe,
however, it is a general problem where ideally the calculator can be
changed. If a decision is made later on to move it out of NewtonNet then
things will (hopefully) be still usable a general path optimization
process and can be added along with the static, relax, freq, and thermo
jobs. But, right now, NewtonNet seems like an OK place for this to be
put in.

 <<

### Checklist

- [ ] I have read the ["Guidelines"
section](https://quantum-accelerators.github.io/quacc/dev/contributing.html#guidelines)
of the contributing guide. Don't lie! 😉
- [ ] My PR is on a custom branch and is _not_ named `main`.
- [ ] I have added relevant, comprehensive [unit
tests](https://quantum-accelerators.github.io/quacc/dev/contributing.html#unit-tests).

### Notes

- Your PR will likely not be merged without proper and thorough tests.
- If you are an external contributor, you will see a comment from
[@buildbot-princeton](https://github.com/buildbot-princeton). This is
solely for the maintainers.
- When your code is ready for review, ping one of the [active
maintainers](https://quantum-accelerators.github.io/quacc/about/contributors.html#active-maintainers).

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Andrew S. Rosen <asrosen93@gmail.com>