BREAKING: Modify AtomSchema
Previously, the Structure (or Molecule) metadata was stored in the root level of the results dictionary. This information is now stored in `structure_metadata` or `molecule_metadata` keys, respectively.

Charge and spin multiplicity information is *not* stored in the `molecules_metadata` entry anymore since this information is stored in the calculator parameters. This greatly simplifies the codebase.