BREAKING: Modify AtomSchema by putting root level structure/molecule metadata one level deeper (#2708)
Previously, the Structure (or Molecule) metadata was stored in the root
level of the results dictionary. This information is now stored in
`structure_metadata` or `molecule_metadata` keys, respectively.

Charge and spin multiplicity information is *not* stored in the
`molecules_metadata` entry anymore since this information is stored in
the calculator parameters. This greatly simplifies the codebase.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>