1.0.0: Update VASP input sets (#2883)
## Summary of Changes

This PR overhauls the input sets in preparation for v1.0.0, to be
released by the end of August. This involves **SIGNIFICANT BREAKING
CHANGES** to the default VASP input sets.

- The default `BulkSet.yaml` will be removed and replaced with
`DefaultSetPBE.yaml`. The two are extremely similar. Updates include:
ALGO = Fast --> ALGO = Normal, GGA_COMPAT = False, and changing the
default Yb and Eu pseudopotentials from _2 to _3.
- The `SlabSet.yaml` will be removed and replaced with `SlabSetPBE.yaml`
to ensure internal consistency with the base `DefaultSetPBE.yaml`.
- The `VASP_PRESET_MAG_DEFAULT` setting has been changed from a default
of 1.0 to 0.5 to prevent accidental convergence to erroneous radical
states, particularly in molecular systems.

I debated where we should change the default k-point grids to be based
on `KSPACING`, but it proved to be more trouble than it was worth, and
the old method is generally sufficient for most cases.

At some point after v1.0.0 is released, we will remove or replace the
`RosenSet` YAMLs that are only used internally since now the base
`DefaultSet` YAMLs are basically the same but faster.